Move field texture to utilities as it was written to be a generic texture method. Factor out renaming and modification of threshold coordinate for shower condition cubes for re-use in modifying the UK model texture fields to make them blendable.
